+++ packages/vscode/generic-essential/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 - Initial release of generic-essential extension pack for vscode
-- 17 required extensions
+- 18 required extensions
 - 1 optional extensions
 
 ### Extensions Included
@@ -16,6 +16,7 @@
 - [Bookmarks](https://marketplace.visualstudio.com/items?itemName&#x3D;alefragnani.bookmarks) - Mark lines and jump to them
 - [Code Spell Checker](https://marketplace.visualstudio.com/items?itemName&#x3D;streetsidesoftware.code-spell-checker) - Spelling checker for source code
 - [DotENV](https://marketplace.visualstudio.com/items?itemName&#x3D;mikestead.dotenv) - Support for dotenv file syntax
+- [Error Lens](https://marketplace.visualstudio.com/items?itemName&#x3D;usernamehw.errorlens) - Improve highlighting of errors, warnings and other language diagnostics
 - [Path Intellisense](https://marketplace.visualstudio.com/items?itemName&#x3D;christian-kohler.path-intellisense) - Visual Studio Code plugin that autocompletes filenames
 - [TODO Tree](https://marketplace.visualstudio.com/items?itemName&#x3D;gruntfuggly.todo-tree) - Show TODO, FIXME, etc. comment tags in a tree view
 - [Version Lens](https://marketplace.visualstudio.com/items?itemName&#x3D;pflannery.vscode-versionlens) - Shows the latest version for each package using code lens
@@ -23,7 +24,7 @@
 - [Markdown All in One](https://marketplace.visualstudio.com/items?itemName&#x3D;yzhang.markdown-all-in-one) - All you need to write Markdown (keyboard shortcuts, table of contents, auto preview and more)
 - [markdownlint](https://marketplace.visualstudio.com/items?itemName&#x3D;davidanson.vscode-markdownlint) - Markdown linting and style checking for Visual Studio Code
 - [Markdown Table Prettifier](https://marketplace.visualstudio.com/items?itemName&#x3D;darkriszty.markdown-table-prettify) - Transforms markdown tables to be more readable
-- [Markdown Preview Mermaid Support](https://marketplace.visualstudio.com/items?itemName&#x3D;yzhang.markdown-all-in-one) - Adds Mermaid diagram and flowchart support to VS Code&#x27;s builtin markdown preview
+- [Markdown Preview Mermaid Support](https://marketplace.visualstudio.com/items?itemName&#x3D;bierner.markdown-mermai) - Adds Mermaid diagram and flowchart support to VS Code&#x27;s builtin markdown preview
 - [Mermaid Markdown Syntax Highlighting](https://marketplace.visualstudio.com/items?itemName&#x3D;bpruitt-goddard.mermaid-markdown-syntax-highlighting) - Markdown syntax support for the Mermaid charting language
 - [Even Better TOML](https://marketplace.visualstudio.com/items?itemName&#x3D;tamasfe.even-better-toml) - Fully-featured TOML support with syntax highlighting, validation, formatting, and schema support
 - [YAML](https://marketplace.visualstudio.com/items?itemName&#x3D;redhat.vscode-yaml) - YAML language support with schema validation