+++ packages/vscodium/godot/CHANGELOG.md
@@ -17,7 +17,7 @@
 - [Godot Files](https://open-vsx.org/extension/alfish/godot-files) - Syntax highlighting for Godot files (.tres, .tscn, .godot, etc.)
 - [Godot Shaders](https://open-vsx.org/extension/arkii/godot-shaders) - Syntax highlighting and support for Godot shader files
 - [Godot Snippets](https://open-vsx.org/extension/razoric/gdscript-snippets) - Code snippets for GDScript and Godot development
-- [Error Lens](https://open-vsx.org/extension/usernamehw/errorlens) - Improve highlighting of errors, warnings and other language diagnostics
+- []() - 
 - [TODO Highlight](https://open-vsx.org/extension/wayou/vscode-todo-highlight) - Highlight TODOs, FIXMEs, and other annotations in your code
 - [Rainbow Brackets](https://open-vsx.org/extension/2gua/rainbow-brackets) - Provide rainbow colors for the round brackets, the square brackets and the squiggly brackets
-- [Better Comments](https://open-vsx.org/extension/aaron-bond/better-comments) - Improve your code commenting by annotating with alert, informational, TODOs, and more
+- []() -