+++ packages/vscodium/typescript/CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Extensions Included
 
 #### Required Extensions
-- [TypeScript Importer](https://open-vsx.org/extension/ms-vscode/vscode-typescript-next) - Automatically searches for TypeScript definitions
-- [TypeScript Snippets](https://open-vsx.org/extension/ms-vscode/vscode-typescript-next) - Code snippets for TypeScript development
-- [TypeScript Debugger](https://open-vsx.org/extension/ms-vscode/js-debug) - Built-in TypeScript/JavaScript debugger for VS Code
+- []() - 
+- []() - 
+- []() -